refactor: improve rocket crate structure
