+++ rocket/src/lib.rs
@@ -4,9 +4,11 @@ extern crate serde;
 extern crate serde_json;
 
 mod error;
-mod query;
-mod respond;
+mod fairing;
 
-pub use self::error::ErrorHandler;
-pub use self::query::*;
-pub use self::respond::*;
+pub mod request;
+pub mod response;
+
+pub use self::fairing::JsonApiFairing;
+pub use self::request::*;
+pub use self::response::*;
